Move `GpioRegisters` to `gpio` module
